feat(ref):[TRI-XXX] - move Digital Twin Registry api client
