+++ irs-api/src/main/java/net/catenax/irs/twinregistry/DigitalTwinRegistryFacade.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package net.catenax.irs.aaswrapper.registry.domain;
+package net.catenax.irs.twinregistry;
 
 import java.util.Collections;
 import java.util.List;
