+++ irs-api/src/test/java/net/catenax/irs/twinregistry/DigitalTwinRegistryExponentialRetryTest.java
@@ -1,4 +1,4 @@
-package net.catenax.irs.aaswrapper.registry.domain;
+package net.catenax.irs.twinregistry;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
